refactor: contract-control to deploy-script-support
