+++ packages/deploy-script-support/package.json
@@ -39,18 +39,23 @@
     "@agoric/import-manager": "workspace:*",
     "@agoric/internal": "workspace:*",
     "@agoric/notifier": "workspace:*",
+    "@agoric/orchestration": "workspace:*",
     "@agoric/pola-io": "workspace:*",
     "@agoric/portfolio-api": "workspace:*",
     "@agoric/smart-wallet": "workspace:*",
     "@agoric/store": "workspace:*",
     "@agoric/time": "workspace:*",
     "@agoric/zoe": "workspace:*",
+    "@agoric/zone": "workspace:*",
     "@endo/base64": "^1.0.12",
     "@endo/bundle-source": "^4.1.2",
     "@endo/errors": "^1.2.13",
+    "@endo/eventual-send": "^1.3.4",
     "@endo/far": "^1.1.14",
     "@endo/marshal": "^1.8.0",
     "@endo/nat": "^5.1.3",
+    "@endo/pass-style": "^1.6.3",
+    "@endo/patterns": "^1.7.0",
     "@endo/promise-kit": "^1.1.13",
     "@endo/zip": "^1.0.11"
   },
