+++ packages/portfolio-deploy/src/access-token-setup.core.js
@@ -15,7 +15,7 @@ const trace = makeTracer('PoC-CE', true);
 /**
  * @import {ERemote} from '@agoric/internal';
  * @import {Board} from '@agoric/vats';
- * @import {AttenuatedDepositPowers} from './attenuated-deposit.core.js';
+ * @import {AttenuatedDepositPowers} from '@agoric/deploy-script-support/src/control/attenuated-deposit.core.js';
  * @import {StorageNode} from '@agoric/internal/src/lib-chainStorage.js';
  * @import {Brand} from '@agoric/ertp';
  * @import {IssuerKit} from '@agoric/ertp';
