+++ packages/portfolio-deploy/src/vstorage-prune.core.js
@@ -5,7 +5,7 @@
  */
 
 /**
- * @import {ChainStoragePresent} from './chain-info.core.js'
+ * @import {ChainStoragePresent} from '@agoric/deploy-script-support/src/control/chain-info.core.js'
  * @import {BootstrapPowers} from '@agoric/vats/src/core/types.js';
  */
