+++ packages/portfolio-deploy/test/portfolio.test.ts
@@ -305,7 +305,9 @@ test.serial('delegate control', async t => {
   // This used to be a single builder but has now been split up
   const materials = Promise.all([
     buildProposal('@aglocal/portfolio-deploy/src/postal-service.build.js'),
-    buildProposal('@aglocal/portfolio-deploy/src/contract-control.build.js'),
+    buildProposal(
+      '@agoric/deploy-script-support/src/control/contract-control.build.js',
+    ),
     buildProposal('@aglocal/portfolio-deploy/src/get-upgrade-kit.build.js'),
     buildProposal('@aglocal/portfolio-deploy/src/portfolio-control.build.js', [
       '--ymaxControlAddress',
